Remove compilation warnings

+++ src/Payload.hs
@@ -70,7 +70,7 @@ parseRequestInput payload = do
 
 doubleSize :: [Word8] -> [Int]
 doubleSize content = do
-  let indexed = zip [0 ..] $ map fromIntegral $ content
+  let indexed = zip [0::Int ..] $ map fromIntegral $ content
   let weak = map snd $ filter (odd . fst) indexed
   let strong = map snd $ filter (even . fst) indexed
   map (\a -> ((fst a) * 256) + (snd a)) $ zip strong weak
